chore: reorg and use Zustand for app state
* reorganized files into subdirs
* broke things down into more specific files
* pull global state out into Zustand
