+++ package.json
@@ -27,7 +27,8 @@
     "pretty-bytes": "^6.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "vaul": "^0.9.1"
+    "vaul": "^0.9.1",
+    "zustand": "^4.5.5"
   },
   "devDependencies": {
     "@types/lodash": "^4.17.7",
