+++ src/App.tsx
@@ -1,23 +1,17 @@
 import CssBaseline from '@mui/material/CssBaseline';
-import Dialog from '@mui/material/Dialog';
 import React from 'react';
-import useMediaQuery from '@mui/material/useMediaQuery';
-import { useTheme } from '@mui/material/styles';
 
 import AppBar from './AppBar';
-import Map from './Map';
-import Packs from './Packs'
+import Map from './Map/Map';
+import PacksDialog from './PacksDialog/PacksDialog';
 
 import './App.css'
-import {usePackStore} from './PackStore';
+import {usePackStore} from './packs/usePackStore';
+import { useSetCurrentPackId } from './useAppStore';
 
 function App() {
-  const [currentPackId, setCurrentPackId] = React.useState<string | null>(null);
-  const [mapType, setMapType] = React.useState<'rocks' | 'water'>('rocks');
-  const [packsModalShown, setPacksModalShown] = React.useState(false);
+  const setCurrentPackId = useSetCurrentPackId();
   const packStore = usePackStore();
-  const theme = useTheme();
-  const isSmall = useMediaQuery(theme.breakpoints.down('md'));
 
   React.useEffect(( ) => {
     async function getCurrentPackId() {
@@ -29,35 +23,14 @@ function App() {
       }
     }
     if (packStore) getCurrentPackId().catch(e => console.error('Failed to get current pack ID', e));
-  }, [packStore]);
+  }, [packStore, setCurrentPackId]);
 
   return (
     <>
       <CssBaseline />
-      <AppBar
-        currentPackId={currentPackId}
-        mapType={mapType}
-        setMapType={setMapType}
-        showPacksModal={() => setPacksModalShown(true)}
-      />
-      <Map
-        mapType={mapType}
-        currentPackId={currentPackId}
-        showPacksModal={() => setPacksModalShown(true)}
-      />
-      <Dialog
-        open={packsModalShown}
-        fullScreen={isSmall}
-        fullWidth
-        maxWidth="lg"
-        onClose={( ) => setPacksModalShown(false)}
-      >
-        <Packs
-          currentPackId={currentPackId}
-          onChoose={setCurrentPackId}
-          onClose={() => setPacksModalShown(false)}
-        />
-      </Dialog>
+      <AppBar />
+      <Map />
+      <PacksDialog />
     </>
   )
 }
