+++ src/AppBar.tsx
@@ -9,15 +9,9 @@ import Typography from '@mui/material/Typography';
 
 import Drawer from './Drawer'
 
-interface Props {
-  currentPackId: string | null;
-  mapType: 'rocks' | 'water';
-  setMapType: (newMapType: 'rocks' | 'water') => void;
-  showPacksModal: ( ) => void;
-}
-
-export default function UnderfootAppBar({ mapType, setMapType, showPacksModal }: Props) {
+export default function UnderfootAppBar() {
   const [drawerOpen, setDrawerOpen] = React.useState(false);
+
   const handleDrawerToggle = () => {
     setDrawerOpen(prevState => !prevState);
   };
@@ -26,10 +20,7 @@ export default function UnderfootAppBar({ mapType, setMapType, showPacksModal }:
     <>
       <Drawer
         drawerOpen={drawerOpen}
-        mapType={mapType}
         setDrawerOpen={setDrawerOpen}
-        setMapType={setMapType}
-        showPacksModal={showPacksModal}
       />
       <Box sx={{ display: 'flex' }}>
         <AppBar component="nav">
